+++ java/com/turbovnc/rfb/SecTypesParameter.java
@@ -71,7 +71,7 @@ public final class SecTypesParameter extends VoidParameter {
       set(defValue);
 
     for (Iterator<Integer> i = enabledSecTypes.iterator(); i.hasNext();) {
-      int refType = (Integer)i.next();
+      int refType = i.next();
       if (refType >= 0x100) {
         result.add(RFB.SECTYPE_VENCRYPT);
         break;
@@ -80,7 +80,7 @@ public final class SecTypesParameter extends VoidParameter {
     result.add(RFB.SECTYPE_TIGHT);
     result.add(RFB.SECTYPE_TLS);
     for (Iterator<Integer> i = enabledSecTypes.iterator(); i.hasNext();) {
-      int refType = (Integer)i.next();
+      int refType = i.next();
       if (refType < 0x100 && refType != RFB.SECTYPE_TIGHT &&
           refType != RFB.SECTYPE_TLS)
         result.add(refType);
@@ -96,7 +96,7 @@ public final class SecTypesParameter extends VoidParameter {
       set(defValue);
 
     for (Iterator<Integer> i = enabledSecTypes.iterator(); i.hasNext();) {
-      int refType = (Integer)i.next();
+      int refType = i.next();
       if (refType != RFB.SECTYPE_VENCRYPT && refType != RFB.SECTYPE_UNIX_LOGIN)
         /* ^^ Do not include VeNCrypt to avoid loops */
         result.add(refType);
@@ -112,7 +112,7 @@ public final class SecTypesParameter extends VoidParameter {
       set(defValue);
 
     for (Iterator<Integer> i = enabledSecTypes.iterator(); i.hasNext();) {
-      int refType = (Integer)i.next();
+      int refType = i.next();
       if (refType < 0x100 && refType != RFB.SECTYPE_VENCRYPT &&
           refType != RFB.SECTYPE_TIGHT && refType != RFB.SECTYPE_TLS)
         result.add(refType);
@@ -126,7 +126,7 @@ public final class SecTypesParameter extends VoidParameter {
       set(defValue);
 
     for (Iterator<Integer> i = enabledSecTypes.iterator(); i.hasNext();)
-      if ((Integer)i.next() == secType)
+      if (i.next() == secType)
         return;
 
     if (isAllowed(secType))
@@ -169,7 +169,7 @@ public final class SecTypesParameter extends VoidParameter {
       return true;
 
     for (i = enabledSecTypes.iterator(); i.hasNext();)
-      if ((Integer)i.next() == secType)
+      if (i.next() == secType)
         return true;
     if (secType == RFB.SECTYPE_VENCRYPT)
       return true;
@@ -204,7 +204,7 @@ public final class SecTypesParameter extends VoidParameter {
   private String getStr(List<Integer> secTypes) {
     StringBuilder sb = new StringBuilder();
     for (Iterator<Integer> i = secTypes.iterator(); i.hasNext();) {
-      int refType = ((Integer)i.next());
+      int refType = i.next();
       if (refType != RFB.SECTYPE_VENCRYPT && refType != RFB.SECTYPE_TIGHT &&
           refType != RFB.SECTYPE_TLS) {
         sb.append(RFB.secTypeName(refType));
